backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
bleach==6.0.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.3
django-cors-headers==4.0.0
django-extensions==3.2.1
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-api-logger==1.1.12
frozenlist==1.3.3
idna==3.4
install==1.3.5
multidict==6.0.4
openai==0.27.7
Pillow==9.5.0
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tqdm==4.65.0
urllib3==2.0.2
webencodings==0.5.1
yarl==1.9.2
